refactor(cmd): extract builder family (catalog, forge, action, curator, investigate, gitops, server) to internal/app
